Make curves smoother by plotting on distance
The x axis of the graph is distance, so we will get much smoother lines
by choosing plot points that are evenly distributed by distance, rather
than time. The only caveat for this method is that the projectile can
obviously only go forwards (albeit not really a problem for this use)!

Note that this only affects projectiles with extreme deceleration.

Test case:
	Velocity = 30 (m/s)
	Angle = 0 (degrees)
	Height = 1.6 (m)
	Mass = 1 (g)
	Diameter = 30 (mm)
	Drag Coefficient = 5
	Gravitational Acceleration = 9.81 (m/s)
	Fluid Density = 1.29 (kg/m^3)